+++ source/Messaging.Infrastructure/OutgoingMessages/Requesting/MessageStorage.cs
@@ -17,6 +17,7 @@ using System.IO;
 using System.Threading.Tasks;
 using Energinet.DataHub.MessageHub.Client.Storage;
 using Messaging.Application.OutgoingMessages;
+using Messaging.Application.OutgoingMessages.Requesting;
 
 namespace Messaging.Infrastructure.OutgoingMessages.Requesting;
