+++ source/Messaging.IntegrationTests/TestDoubles/MessageStorageSpy.cs
@@ -16,6 +16,7 @@ using System;
 using System.IO;
 using System.Threading.Tasks;
 using Messaging.Application.OutgoingMessages;
+using Messaging.Application.OutgoingMessages.Requesting;
 using Xunit;
 
 namespace Messaging.IntegrationTests.TestDoubles;
